moved test to 03 and improved slightly speed
